feat: controllers acoplados
