+++ server/src/routes/memoriesRoutes.ts
@@ -1,0 +1,8 @@
+import { FastifyInstance } from 'fastify';
+
+import { memoriesController } from '../contollers/memoriesController';
+
+export async function memoriesRoutes(app: FastifyInstance): Promise<void> {
+  app.get('/memories', memoriesController.getAll);
+  app.get('/memories/:id', memoriesController.getById);
+}
